requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic-core==2.27.1
pypdf==5.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
